Add completions for keywords, constants, properties, commands, and functions.
Note that the function parameter lists have not been filled in. All have exactly one parameter. This is left as something to do in the future.

LiveCode is interesting in that something like `the color names` is really a function (`colornames()`). The function version is used in completion. You can quickly delete `()` after completing it you are using the property. Just press TAB and then delete.